# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SRI - Release Go Live Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /perfis:
    get:
      operationId: listPerfis
      tags: [perfis]
      summary: List access profiles (admin only)
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Perfil"
    post:
      operationId: createPerfil
      tags: [perfis]
      summary: Create an access profile (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerfilInput"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Perfil"
        "409":
          description: Profile key already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /perfis/{id}:
    patch:
      operationId: updatePerfil
      tags: [perfis]
      summary: Update an access profile (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerfilUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Perfil"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deletePerfil
      tags: [perfis]
      summary: Delete a custom access profile (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: System profile cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /perfis/{id}/permissoes:
    get:
      operationId: getPerfilPermissoes
      tags: [perfis]
      summary: Get the permission grants of a profile (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permission grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permissao"
    put:
      operationId: setPerfilPermissoes
      tags: [perfis]
      summary: Replace the permission grants of a profile (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerfilPermissoesInput"
      responses:
        "200":
          description: Updated permission grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permissao"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /objetos:
    get:
      operationId: listObjetos
      tags: [objetos]
      summary: List system objects/resources (admin only)
      responses:
        "200":
          description: List of objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Objeto"
    post:
      operationId: createObjeto
      tags: [objetos]
      summary: Create a system object (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjetoInput"
      responses:
        "201":
          description: Created object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objeto"
        "409":
          description: Object key already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /objetos/{id}:
    patch:
      operationId: updateObjeto
      tags: [objetos]
      summary: Update a system object (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjetoUpdate"
      responses:
        "200":
          description: Updated object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objeto"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteObjeto
      tags: [objetos]
      summary: Delete a custom object (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: System object cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /lookups:
    get:
      operationId: listLookups
      tags: [lookups]
      summary: List domain lookup values
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of lookup values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lookup"
    post:
      operationId: createLookup
      tags: [lookups]
      summary: Create a lookup value (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupInput"
      responses:
        "201":
          description: Created lookup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lookup"
  /lookups/{id}:
    patch:
      operationId: updateLookup
      tags: [lookups]
      summary: Update a lookup value (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupUpdate"
      responses:
        "200":
          description: Updated lookup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lookup"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteLookup
      tags: [lookups]
      summary: Delete a lookup value (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Perfil:
      type: object
      required: [id, chave, nome, sistema]
      properties:
        id:
          type: integer
        chave:
          type: string
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        sistema:
          type: boolean
          description: System profiles cannot be deleted.
        createdAt:
          type: ["string", "null"]
    PerfilInput:
      type: object
      required: [nome]
      properties:
        chave:
          type: string
          description: Optional unique key; derived from nome when omitted.
        nome:
          type: string
          minLength: 1
        descricao:
          type: string
    PerfilUpdate:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
    Objeto:
      type: object
      required: [id, chave, nome, acoes, sistema]
      properties:
        id:
          type: integer
        chave:
          type: string
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        acoes:
          type: array
          items:
            type: string
        sistema:
          type: boolean
          description: System objects cannot be deleted.
        createdAt:
          type: ["string", "null"]
    ObjetoInput:
      type: object
      required: [chave, nome, acoes]
      properties:
        chave:
          type: string
          minLength: 1
        nome:
          type: string
          minLength: 1
        descricao:
          type: string
        acoes:
          type: array
          items:
            type: string
    ObjetoUpdate:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        acoes:
          type: array
          items:
            type: string
    Permissao:
      type: object
      required: [objetoId, acao]
      properties:
        objetoId:
          type: integer
        acao:
          type: string
    PerfilPermissoesInput:
      type: object
      required: [permissoes]
      properties:
        permissoes:
          type: array
          items:
            $ref: "#/components/schemas/Permissao"
    Lookup:
      type: object
      required: [id, category, value]
      properties:
        id:
          type: integer
        category:
          type: string
        value:
          type: string
        ordem:
          type: ["integer", "null"]
        createdAt:
          type: ["string", "null"]
    LookupInput:
      type: object
      required: [category, value]
      properties:
        category:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        ordem:
          type: integer
    LookupUpdate:
      type: object
      properties:
        category:
          type: string
        value:
          type: string
        ordem:
          type: integer
...