# lib/api-zod/src/generated/types/objeto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SRI - Release Go Live Control API
 * OpenAPI spec version: 0.1.0
 */

export interface Objeto {
  id: number;
  chave: string;
  nome: string;
  /** @nullable */
  descricao?: string | null;
  acoes: string[];
  /** System objects cannot be deleted. */
  sistema: boolean;
  /** @nullable */
  createdAt?: string | null;
}