# lib/api-zod/src/generated/types/objetoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SRI - Release Go Live Control API
 * OpenAPI spec version: 0.1.0
 */

export interface ObjetoInput {
  /** @minLength 1 */
  chave: string;
  /** @minLength 1 */
  nome: string;
  descricao?: string;
  acoes: string[];
}