# lib/api-zod/src/generated/types/perfilInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SRI - Release Go Live Control API
 * OpenAPI spec version: 0.1.0
 */

export interface PerfilInput {
  /** Optional unique key; derived from nome when omitted. */
  chave?: string;
  /** @minLength 1 */
  nome: string;
  descricao?: string;
}