# lib/api-zod/src/generated/types/perfilPermissoesInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SRI - Release Go Live Control API
 * OpenAPI spec version: 0.1.0
 */
import type { Permissao } from './permissao';

export interface PerfilPermissoesInput {
  permissoes: Permissao[];
}